Refacto: Dto 이름변경 및 Querydsl 적용. 아직 제대로 안돌아가는상태.

+++ src/main/java/team7/hrbank/domain/employee/repository/EmployeeRepository.java
@@ -9,5 +9,5 @@ public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
     // 부서에 소속된 직원수 확인
     @Query("SELECT COUNT(e) FROM Employee e WHERE e.department.id = :departmentId")
-    Integer countEmployeesByDepartmentId(@Param("departmentId") Long departmentId);
+    Long countEmployeesByDepartmentId(@Param("departmentId") Long departmentId);
 }
